+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Improve output with exit code and stderr
 - Terminate commands if not finished after a timeout (default 20s)
 - Show a progress notification display with a cancel button for long running commands.
+- Automatically open a preview on file open if the file type is configured (with customizable position and toggle settings)
 
 ## [0.0.3] - 2026-05-15